+++ src/ExceptionHandling/ChainExceptionHandler.php
@@ -15,7 +15,6 @@ declare(strict_types = 1);
 namespace SensioLabs\RichModelForms\ExceptionHandling;
 
 use Symfony\Component\Form\FormConfigInterface;
-use Symfony\Component\Form\FormError;
 
 /**
  * Delegates execution to a list of handlers, stopping after the first handler that transformed the exception.
@@ -40,7 +39,7 @@ final class ChainExceptionHandler implements ExceptionHandlerInterface
         $this->exceptionHandlers = $exceptionHandlers;
     }
 
-    public function getError(FormConfigInterface $formConfig, $data, \Throwable $e): ?FormError
+    public function getError(FormConfigInterface $formConfig, $data, \Throwable $e): ?Error
     {
         foreach ($this->exceptionHandlers as $exceptionHandler) {
             if (null !== $error = $exceptionHandler->getError($formConfig, $data, $e)) {